// src/lib/server/schemas/rsstream.ts
export const feedItemDataSchema = z.object({
  title: z.string(),
  content: z.string(),
  author: z.string(),
  permalink: z.string(),
  guid: z.string(),
  timestamp: z.number(),
  url: z.string(),
  enclosure: z.object({
    photo: z.record(z.string()).optional(),
    audio: z.record(z.boolean()).optional(),
    video: z.record(z.boolean()).optional(),
  }).optional(),
  category: z.record(z.boolean()).optional(),
})
// ...
export type FeedItemData = z.infer<typeof feedItemDataSchema>